Prune low-signal tests and halve covered suite runtime
