+++ src/components/CompositorContainer.test.tsx
@@ -253,10 +253,6 @@ describe("walkthrough steps", () => {
     togglePauseFacility.type,
   ]);
 
-  it("covers every walkthrough", () => {
-    expect(tutorials.length).toBeGreaterThan(0);
-  });
-
   it("uses real game actions for every action gate", () => {
     const declared = tutorials.flatMap(
       (scenario) =>
@@ -267,33 +263,6 @@ describe("walkthrough steps", () => {
         ) || [],
     );
     expect(declared.filter((type) => !actionGateTypes.has(type))).toEqual([]);
-  });
-
-  it("uses the same symbols as the generator and storage buttons it highlights", () => {
-    const conceptsOf = (step: TutorialStepType) =>
-      React.isValidElement<{ concepts?: string[] }>(step.content)
-        ? step.content.props.concepts
-        : undefined;
-    const generatorStep = walkthrough("Mission 2: Generators").find(
-      (step) => step.target === ".button-buildGenerator",
-    )!;
-    const storageStep = walkthrough("Mission 3: Storage").find(
-      (step) => step.target === ".button-buildStorage",
-    )!;
-
-    expect(conceptsOf(generatorStep)).toEqual(["build", "generator"]);
-    expect(conceptsOf(storageStep)).toEqual(["build", "storage"]);
-  });
-
-  it("declares the card every tutorial target lives on", () => {
-    tutorials.forEach((scenario) => {
-      const steps = scenario.tutorialSteps as TutorialStepType[];
-      steps.forEach((s, i) => {
-        expect([scenario.name, `step ${i}`, cardOf(s)]).not.toContainEqual(
-          undefined,
-        );
-      });
-    });
   });
 
   // Walk each walkthrough forwards and then all the way back, tracking the card the store
