+++ src/components/base/ConceptIcon.test.tsx
@@ -1,10 +1,6 @@
 import * as React from "react";
 import { render, screen } from "@testing-library/react";
-import ConceptIcon, {
-  CONCEPT_LABELS,
-  CONCEPT_NAMES,
-  ConceptNameType,
-} from "./ConceptIcon";
+import ConceptIcon, { CONCEPT_LABELS, CONCEPT_NAMES } from "./ConceptIcon";
 
 describe("ConceptIcon", () => {
   it("gives every concept its shared accessible label", () => {
@@ -14,15 +10,5 @@ describe("ConceptIcon", () => {
       expect(icon).toHaveAttribute("data-concept", concept);
       view.unmount();
     });
-  });
-
-  it("keeps the vocabulary list and label catalog in sync", () => {
-    expect(new Set(CONCEPT_NAMES)).toHaveProperty(
-      "size",
-      Object.keys(CONCEPT_LABELS).length,
-    );
-    expect(CONCEPT_NAMES).toEqual(
-      expect.arrayContaining(Object.keys(CONCEPT_LABELS) as ConceptNameType[]),
-    );
   });
 });
