+++ src/components/base/TutorialPrompt.test.tsx
@@ -3,19 +3,6 @@ import { render, screen } from "@testing-library/react";
 import TutorialPrompt from "./TutorialPrompt";
 
 describe("TutorialPrompt", () => {
-  it("keeps the instruction to one concise line", () => {
-    render(
-      <TutorialPrompt
-        concepts={["build", "generator"]}
-        text="Open the generator shop."
-      />,
-    );
-
-    expect(screen.getByText("Open the generator shop.")).toBeInTheDocument();
-    expect(screen.queryByText("Do this to continue:")).toBeNull();
-    expect(screen.queryByRole("status")).toBeNull();
-  });
-
   it("gives the concept sequence an accessible summary", () => {
     render(<TutorialPrompt concepts={["supply", "demand"]} />);
