+++ src/components/views/EventLog.test.tsx
@@ -102,42 +102,6 @@ describe("EventLog", () => {
     expect(screen.getByText("Event history")).toBeVisible();
   });
 
-  it("keeps upcoming and past events in labeled sections", () => {
-    render(
-      <EventLog
-        events={[
-          {
-            id: 1,
-            kind: "BUILD",
-            label: "Jan 2024",
-            message: "Solar project started.",
-          },
-          {
-            id: 2,
-            kind: "CONSTRUCTION",
-            label: "Feb 2024",
-            message: "Solar project finished.",
-          },
-        ]}
-        upcoming={[
-          {
-            key: "next",
-            label: "Mar 2024",
-            message: "Demand will rise.",
-          },
-        ]}
-        onOpen={jest.fn()}
-        onSelect={jest.fn()}
-      />,
-    );
-
-    expect(screen.getByText("Upcoming events")).toBeVisible();
-    expect(screen.getByText("Event history")).toBeVisible();
-    expect(
-      screen.queryByText(/recorded after they happen/i),
-    ).not.toBeInTheDocument();
-  });
-
   it("filters event history into useful event groups", async () => {
     const user = userEvent.setup();
     render(
