+++ src/components/views/MainMenu.test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { render, screen, within } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import MainMenu, { Props } from "./MainMenu";
 
@@ -18,15 +18,6 @@ function props(overrides: Partial<Props> = {}): Props {
 }
 
 describe("MainMenu", () => {
-  it("explains the game without specialist language", () => {
-    render(<MainMenu {...props()} />);
-
-    expect(
-      screen.getByText(/Keep the lights on. Build a cleaner grid/i),
-    ).toHaveClass("gameSubtitle", "MuiTypography-body1");
-    expect(screen.queryByText(/no energy or gaming experience/i)).toBeNull();
-  });
-
   it("starts a new game from the primary play action", async () => {
     const onStart = jest.fn();
     const user = userEvent.setup();
@@ -35,56 +26,6 @@ describe("MainMenu", () => {
     await user.click(screen.getByRole("button", { name: "Start playing" }));
     expect(onStart).toHaveBeenCalled();
     expect(screen.queryByText("Continue")).not.toBeInTheDocument();
-  });
-
-  it("only advertises account-free play before sign-in", () => {
-    const { rerender } = render(<MainMenu {...props({ uid: undefined })} />);
-    const accountActions = screen.getByRole("region", {
-      name: "Account actions",
-    });
-    expect(
-      within(accountActions).getByRole("button", { name: "Sign in" }),
-    ).toBeInTheDocument();
-    expect(
-      within(accountActions).getByText("Free · no sign-up needed"),
-    ).toBeInTheDocument();
-
-    rerender(<MainMenu {...props({ uid: "player" })} />);
-    expect(
-      screen.queryByText("Free · no sign-up needed"),
-    ).not.toBeInTheDocument();
-  });
-
-  it("places account actions below the sound and discovery actions", () => {
-    render(
-      <MainMenu
-        {...props({
-          audioEnabled: undefined,
-          hasSavedGame: false,
-          uid: undefined,
-        })}
-      />,
-    );
-
-    const resources = screen.getByRole("navigation", {
-      name: "Game resources",
-    });
-    const account = screen.getByRole("region", { name: "Account actions" });
-    const discovery = screen.getByRole("region", {
-      name: "Discovery actions",
-    });
-    expect(
-      resources.compareDocumentPosition(discovery) &
-        Node.DOCUMENT_POSITION_FOLLOWING,
-    ).toBeTruthy();
-    expect(
-      discovery.compareDocumentPosition(account) &
-        Node.DOCUMENT_POSITION_FOLLOWING,
-    ).toBeTruthy();
-    expect(within(account).getByText("Free · no sign-up needed")).toBeVisible();
-    expect(
-      within(account).getByRole("button", { name: "Sign in" }),
-    ).toBeVisible();
   });
 
   it("prioritizes continuing a save while offering mission selection", () => {
@@ -101,53 +42,6 @@ describe("MainMenu", () => {
     ).toBeInTheDocument();
     expect(screen.getByRole("region", { name: "Primary actions" })).toHaveStyle(
       { gap: "10px" },
-    );
-  });
-
-  it("groups secondary actions into compact rows", () => {
-    render(
-      <MainMenu
-        {...props({
-          audioEnabled: undefined,
-          hasSavedGame: true,
-          uid: undefined,
-        })}
-      />,
-    );
-
-    const primary = screen.getByRole("region", { name: "Primary actions" });
-    const resources = screen.getByRole("navigation", {
-      name: "Game resources",
-    });
-    const discovery = screen.getByRole("region", {
-      name: "Discovery actions",
-    });
-
-    expect(primary).toHaveStyle({ flexDirection: "column" });
-    expect(resources).toHaveStyle({ flexDirection: "row", gap: "6px" });
-    expect(
-      within(resources).queryByRole("button", { name: "Sign in" }),
-    ).not.toBeInTheDocument();
-    expect(discovery).toHaveStyle({ flexDirection: "row", gap: "6px" });
-  });
-
-  it("keeps sharing as a compact footer icon", () => {
-    render(<MainMenu {...props()} />);
-
-    expect(
-      within(screen.getByRole("contentinfo")).getByRole("button", {
-        name: "Share Electrify",
-      }),
-    ).toBeInTheDocument();
-    expect(screen.queryByRole("button", { name: "Share game" })).toBeNull();
-  });
-
-  it("opens the embedded feedback form without exposing an email address", () => {
-    render(<MainMenu {...props()} />);
-
-    expect(screen.getByRole("link", { name: "Send feedback" })).toHaveAttribute(
-      "href",
-      "/about.html#feedback",
     );
   });
 });
