+++ src/components/views/Manual.test.tsx
@@ -31,16 +31,6 @@ describe("Manual", () => {
     clearManualMemory();
   });
 
-  it("pins How to Play above the grouped entries", () => {
-    renderManual();
-    const titles = listedTitles();
-    expect(titles[0]).toContain(MANUAL_ENTRY.HOW_TO_PLAY);
-    // Alphabetical ordering used to open on these two, which tell a new player nothing
-    expect(titles[1]).not.toContain(MANUAL_ENTRY.BTU);
-    expect(screen.getByText("Gameplay")).toBeInTheDocument();
-    expect(screen.getByText("Physics & Units")).toBeInTheDocument();
-  });
-
   // The old filter only looked at children that were plain strings, so any paragraph
   // containing an element (<strong>, a nested list, ...) was invisible to search
   it("finds representative terms inside mixed markup", async () => {
@@ -61,21 +51,6 @@ describe("Manual", () => {
     renderManual();
     await search("LCOE");
     expect(entryHeader(MANUAL_ENTRY.TOTAL_COST_OF_ENERGY)).toBeInTheDocument();
-  });
-
-  it("has entries for the terms the game shows on screen", async () => {
-    renderManual();
-    for (const term of [
-      "capacity factor",
-      "ramp rate",
-      "peaker",
-      "round-trip efficiency",
-      "rates",
-      "carbon fee",
-    ]) {
-      await search(term);
-      expect(listedTitles().length).toBeGreaterThan(0);
-    }
   });
 
   it("includes every shared game symbol in the symbol guide", async () => {
@@ -137,11 +112,6 @@ describe("Manual", () => {
     expect(entryHeader(MANUAL_ENTRY.BTU)).toBeInTheDocument();
   });
 
-  it("has no clear button until there's something to clear", () => {
-    renderManual();
-    expect(screen.queryByLabelText("clear search")).not.toBeInTheDocument();
-  });
-
   it("opens the entry a deep link points at, and nothing else", () => {
     renderManual(MANUAL_ENTRY.TOTAL_COST_OF_ENERGY);
     expect(entryHeader(MANUAL_ENTRY.TOTAL_COST_OF_ENERGY)).toHaveAttribute(
@@ -160,21 +130,6 @@ describe("Manual", () => {
     expect(header).toHaveAttribute("tabindex", "0");
     expect(header).toHaveAttribute("aria-expanded");
     expect(header).toHaveAttribute("aria-controls");
-  });
-
-  it("lists the point values in a column rather than as run-on text", async () => {
-    renderManual();
-    await userEvent.click(entryHeader(MANUAL_ENTRY.SCORE));
-    const row = screen
-      .getAllByRole("row")
-      .find((r: HTMLElement) =>
-        (r.textContent || "").includes(
-          "per TWh of demand not served during blackouts",
-        ),
-      );
-    expect(row).toBeDefined();
-    // The investor penalty, in its own cell rather than run together with the text
-    expect(within(row as HTMLElement).getByText("-8")).toBeInTheDocument();
   });
 
   it("remembers the search term across visits", async () => {
