+++ src/data/Scenarios.test.tsx
@@ -7,15 +7,8 @@ import {
   TUTORIALS,
 } from "./Scenarios";
 import { AppStateType, ScenarioType } from "../Types";
-import { render, screen } from "@testing-library/react";
-import { getScenarioLocation } from "../helpers/Locations";
 
 describe("getScenario", () => {
-  it("finds an authored scenario by id", () => {
-    const authored = SCENARIOS[SCENARIOS.length - 1];
-    expect(getScenario(authored.id)).toBe(authored);
-  });
-
   it("returns the custom scenario for the custom id", () => {
     const custom = {
       ...DEFAULT_CUSTOM_SCENARIO,
@@ -64,25 +57,6 @@ describe("getNextTutorial", () => {
 });
 
 describe("tutorial mission metadata", () => {
-  it("gives every tutorial a mission name, icon, and summary", () => {
-    TUTORIALS.forEach((tutorial, index) => {
-      expect(tutorial.name).toMatch(new RegExp(`^Mission ${index + 1}: `));
-      expect(tutorial.icon).toBeTruthy();
-      expect(tutorial.summary).toBeTruthy();
-    });
-  });
-
-  it("expands the O&M abbreviation in the generator tutorial", () => {
-    const generatorsMission = TUTORIALS.find(
-      (tutorial) => tutorial.name === "Mission 2: Generators",
-    )!;
-    render(generatorsMission.tutorialSteps![1].content);
-
-    expect(screen.getByText(/Compare cost and build time/)).toHaveTextContent(
-      "operations and maintenance (O&M)",
-    );
-  });
-
   it("advances the finances tutorial when the mobile Insights tab opens", () => {
     const finances = TUTORIALS.find(
       (tutorial) => tutorial.name === "Mission 4: Finances",
@@ -105,53 +79,6 @@ describe("tutorial mission metadata", () => {
       expect(capstones[0].target).toBeUndefined();
       expect(capstones[0].hint).toBeTruthy();
     });
-  });
-
-  it("ends the electricity mission after a single one-day challenge", () => {
-    const electricity = TUTORIALS.find(
-      (tutorial) => tutorial.name === "Mission 1: Electricity",
-    )!;
-    const steps = electricity.tutorialSteps!;
-
-    expect(steps).toHaveLength(5);
-    expect(steps[3].advanceOn).toBeDefined();
-    expect(steps[4].advanceOn).toBeUndefined();
-    expect(steps[4].capstone).toBeDefined();
-  });
-});
-
-describe("authored scenario briefings", () => {
-  it("gives every scored scenario a reusable story and stakes", () => {
-    SCENARIOS.filter((scenario) => !scenario.tutorialSteps).forEach(
-      (scenario) => {
-        expect(scenario.briefing).toEqual(
-          expect.objectContaining({
-            tone: expect.any(String),
-            fantasy: expect.any(String),
-            objective: expect.any(String),
-            threat: expect.any(String),
-          }),
-        );
-        expect(scenario.briefing).not.toHaveProperty("constraint");
-      },
-    );
-  });
-
-  it("gives every challenge at least one player-facing browse theme", () => {
-    SCENARIOS.filter((scenario) => !scenario.tutorialSteps).forEach(
-      (scenario) => expect(scenario.themes?.length).toBeGreaterThan(0),
-    );
-  });
-
-  it("uses the three player-facing challenge themes", () => {
-    const themes = new Set(
-      SCENARIOS.filter((scenario) => !scenario.tutorialSteps).flatMap(
-        (scenario) => scenario.themes ?? [],
-      ),
-    );
-    expect(themes).toEqual(
-      new Set(["Extreme weather", "Energy transition", "Rapid growth"]),
-    );
   });
 });
 
@@ -272,29 +199,6 @@ describe("authored starting fleets", () => {
     );
   });
 
-  it("puts each scored scenario's defining facility first", () => {
-    expect(
-      SCENARIOS.filter((scenario) => !scenario.tutorialSteps).map(
-        (scenario) => [
-          scenario.name,
-          scenario.facilities[0].fuel || scenario.facilities[0].name,
-        ],
-      ),
-    ).toEqual([
-      ["Carbon Fee", "Natural Gas"],
-      ["The Shale Boom", "Coal"],
-      ["Paradise", "Sun"],
-      ["Rise of Renewables", "Uranium"],
-      ["Hurricane Season", "Oil"],
-      ["The End of an Era", "Coal"],
-      ["Data Center Boom", "Natural Gas"],
-      ["Deep Freeze", "Natural Gas"],
-      ["Heatwave + Drought", "Hydro"],
-      ["Sudden Nuclear Shutdown", "Uranium"],
-      ["Wildfire Emergency", "Natural Gas"],
-    ]);
-  });
-
   it("authors distinct heatwave and generation-loss resilience challenges", () => {
     const heatwave = getScenario(108)!;
     const trip = getScenario(110)!;
@@ -330,20 +234,6 @@ describe("authored starting fleets", () => {
         (facility) => facility.label === "Grand Nuclear Unit",
       ),
     ).toMatchObject({ fuel: "Uranium", peakW: 500_000_000 });
-  });
-
-  it("keeps locations in scenario metadata rather than scenario names", () => {
-    expect(
-      [107, 108, 110, 111].map((id) => ({
-        name: getScenario(id)!.name,
-        location: getScenarioLocation(getScenario(id))?.name,
-      })),
-    ).toEqual([
-      { name: "Deep Freeze", location: "Austin, TX" },
-      { name: "Heatwave + Drought", location: "Madrid, Spain" },
-      { name: "Sudden Nuclear Shutdown", location: "Paris, France" },
-      { name: "Wildfire Emergency", location: "Los Angeles, CA" },
-    ]);
   });
 
   it("authors a Los Angeles-only January 2025 wildfire challenge", () => {
